[api] v3 /lots/update. reject updates on validated and accepted lots

+++ web/api/v3/lots/views.py
@@ -362,6 +362,8 @@ def update_lot(request):
     except Exception as e:
         return JsonResponse({'status': 'error', 'message': "Unknown transaction %s" % (tx_id), 'extra': str(e)},
                             status=400)
+    if tx.delivery_status == 'A':
+        return JsonResponse({'status': 'forbidden', 'message': "Tx / Lot already validated and accepted"}, status=400)
     LotV2Error.objects.filter(id=tx.lot.id).delete()
     TransactionError.objects.filter(id=tx.id).delete()
     load_lot(entity, request.user, request.POST.dict(), 'MANUAL', tx)
